Reorganising tests for facet groups
